Priyanshu Ranjan/requirements.txt
streamlit>=1.24
pandas
python-docx
PyPDF2
ollama
sentence-transformers
faiss-cpu
